fix: use SqlLiteQueries (basic)
