+++ .gitignore
@@ -63,3 +63,5 @@ typings/
 # clinic
 *clinic*
 package-lock.json
+
+database.sqlite
