+++ examples/query-filters/README.md
@@ -1,0 +1,7 @@
+## Query filtering example
+
+Uses SqlLite database to showcase graphql filtering capabilities
+
+## Running 
+
+node ./queryFiters.js
