+++ package.json
@@ -23,6 +23,7 @@
     "graphql-tools": "^4.0.4",
     "pre-commit": "^1.2.2",
     "snazzy": "^8.0.0",
+    "sqlite": "^3.0.3",
     "standard": "^12.0.1",
     "tap": "^14.1.6"
   },
